Make Generated import non qualified
I don't know what is supposed to happen if one exports a qualifiedly
imported module, but what does happen is that the names in the exported
module are not (easily) accessible.

A google search got me no results to exports of qualified imports, but
simply dropping qualified made it work for me.